+++ src/sections/ResourceType.vue
@@ -5,6 +5,7 @@ import SectionBorder from '../components/SectionBorder.vue'
 import SectionTitle from '../components/SectionTitle.vue'
 import Output from '../components/Output.vue'
 import FormLabel from '../components/FormLabel.vue'
+import TwoColumn from '../components/TwoColumn.vue'
 
 const emit = defineEmits(['update:resourceType'])
 
@@ -27,8 +28,8 @@ watch(
 <template>
   <SectionBorder>
     <SectionTitle anchor="resource-type" title="Resource Type" />
-    <div class="flex">
-      <form class="w-1/2 pr-2 flex flex-col">
+    <TwoColumn>
+      <template v-slot:left>
         <FormLabel v-for="type in resourceTypes" :key="type">
           <input
             type="radio"
@@ -39,10 +40,10 @@ watch(
           />
           {{ type }}
         </FormLabel>
-      </form>
-      <div class="w-1/2 pl-2 flex flex-col">
+      </template>
+      <template v-slot:right>
         <Output :data="resourceType"></Output>
-      </div>
-    </div>
+      </template>
+    </TwoColumn>
   </SectionBorder>
 </template>